pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jobapply"
version = "0.1.0"
description = "CLI to search jobs (JobSpy), tailor resume & cover letter with LangGraph agents, export JSON + PDFs."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "JobApply contributors" }]
keywords = ["jobs", "langgraph", "langchain", "resume", "cli", "jobspy"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Office/Business",
]
dependencies = [
  "python-jobspy>=1.1.80",
  "langgraph>=0.2.60",
  "langgraph-checkpoint-sqlite>=2.0.0",
  "langchain-core>=0.3.0",
  "langchain-google-genai>=2.0.0",
  "langchain-anthropic>=0.3.0",
  "langchain-ollama>=0.2.0",
  "langchain-openai>=0.2.0",
  "sqlmodel>=0.0.22",
  "pydantic>=2.7",
  "typer>=0.12.0",
  "questionary>=2.0.0",
  "rich>=13.7.0",
  "jinja2>=3.1.0",
  "pypandoc>=1.13",
  "tenacity>=8.2.0",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.6.0",
  "black>=24.0.0",
  "mypy>=1.11.0",
  "types-requests",
]

[project.scripts]
jobapply = "jobapply.cli:main"

[project.urls]
Homepage = "https://github.com/arun2728/jobs-search-and-apply-api"
Repository = "https://github.com/arun2728/jobs-search-and-apply-api"
Issues = "https://github.com/arun2728/jobs-search-and-apply-api/issues"

[tool.hatch.build.targets.wheel]
packages = ["jobapply"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_ignores = true
packages = ["jobapply"]
disallow_untyped_defs = false
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
  "jobspy",
  "langgraph.*",
  "langgraph.checkpoint.sqlite",
  "langchain_core.*",
  "langchain_google_genai.*",
  "langchain_anthropic.*",
  "langchain_ollama.*",
  "langchain_openai.*",
  "pypandoc",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
